+++ src/kernel/IO/ConsoleIn.h
@@ -8,5 +8,5 @@
 class ConsoleIn final : public AbstractFile  {
 
 public:
-	kiv_os::NOS_Error read(char* targetBuffer, uint32_t bytes, uint32_t& bytesRead) override;
+	kiv_os::NOS_Error read(char* targetBuffer, size_t bytes, size_t& bytesRead) override;
 };
